COSBETA-24: create notification from a zip fiel. seperate a job that is called by analyzer.
