+++ cosmetics-web/Gemfile
@@ -11,6 +11,7 @@ gem "mini_magick"
 gem "pg", "~> 1.1"
 gem "puma", "~> 3.0"
 gem "rails", "~> 5.2"
+gem "rubyzip", ">= 1.2.1"
 gem "sass-rails", "~> 5.0"
 gem "slim-rails"
 gem "strong_migrations"
